Organize backend into package
