+++ README.md
@@ -1,1 +1,15 @@
 # InnopolisAlias
+
+This repository contains a small FastAPI backend for an Alias-like game.
+
+## Structure
+
+```
+backend/
+    requirements.txt  # project dependencies
+    app/
+        main.py       # FastAPI application
+        db.py         # database connection
+        routers/      # API and websocket routers
+        models/       # Pydantic models
+```
