+++ backend/app/main.py
@@ -2,8 +2,8 @@ import uvicorn
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 
-from auth import router as auth_router
-from game import router as game_router
+from .routers.auth import router as auth_router
+from .routers.game import router as game_router
 
 app = FastAPI()
 
@@ -19,4 +19,4 @@ app.include_router( game_router, prefix="/api/game", tags=["game"])
 
 
 if __name__ == "__main__":
-    uvicorn.run("main:app", host="127.0.0.1", port=8000, reload=True)
+    uvicorn.run("app.main:app", host="127.0.0.1", port=8000, reload=True)
